add Icon Component to display Icons
I want to use icons but dont want to add jpegs and display them as img. So I created a Icon component which makes use of svg icons.
The icon path for the vector is stored in an IconConstants file so if I wish to change the icon or add more it just take 1 to 2 lines of code